# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Session cookie (browser) or Bearer token (scripts, mobile). Cookie-authenticated POST/PUT/PATCH/DELETE must send X-Rigel-Client.",
        "title": "RigelLedger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/admin/access-requests": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Requests for an account (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pending, approved or rejected",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.AccessRequestDTO"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/access-requests/{requestID}/approve": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Approve a request: the person is invited (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/access-requests/{requestID}/reject": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Reject a request (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Recent sign-in and security events of everyone (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "at most 500",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.AuthEventDTO"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/invitations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Invite a user by username and email (admin)",
                "parameters": [
                    {
                        "description": "who",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.inviteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminUserDTO"
                        }
                    },
                    "409": {
                        "description": "taken, or mail_failed (the invitation is saved; resend it)",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/admin/invitations/{userID}": {
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Withdraw an invitation that was not accepted; deletes that user (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "invited user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/invitations/{userID}/resend": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Mail a fresh invitation link; the old one stops working (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "invited user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/admin/mail": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Outgoing mail settings (admin); the password is stored sealed and never returned",
                "parameters": [
                    {
                        "description": "mail",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminMailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    }
                }
            }
        },
        "/api/admin/mail/test": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Send a test email to yourself (admin)",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "409": {
                        "description": "mail_failed or mail_disabled",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/admin/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "System settings (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Replace registration, security and default settings (admin)",
                "parameters": [
                    {
                        "description": "every field",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminSettingsDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{userID}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Activate, deactivate, promote or demote another user (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "changes",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.adminUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AdminUserDTO"
                        }
                    },
                    "409": {
                        "description": "last_admin",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/auth/invite/{token}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "The invitation behind a link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token from the link",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.InvitationDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Accept an invitation: settings and password, then signed in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token from the link",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "profile and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.acceptInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.signedInResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.loginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "too many failures; see Retry-After",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up (registration = open); a confirmation link is mailed",
                "parameters": [
                    {
                        "description": "sign-up",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "403": {
                        "description": "registration_closed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/auth/request-access": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Ask the administrators for an account (registration = request)",
                "parameters": [
                    {
                        "description": "request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.accessRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/api/auth/verify-link": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Confirm a sign-up link: creates the user and signs them in",
                "parameters": [
                    {
                        "description": "the token from the link",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.tokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.signedInResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Books the user is a member of",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.BookDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Create a book seeded with the personal chart of accounts",
                "parameters": [
                    {
                        "description": "book",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.createBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.BookDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/books/{bookID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "One book and the caller's role in it",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.BookDTO"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Rename, lock, or set IAS 7 choices (owner)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "book",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.updateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.BookDTO"
                        }
                    }
                }
            }
        },
...
        "/api/books/{bookID}/accounts/{accountID}/archive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Archive or restore an account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "archived flag",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.archiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.AccountDTO"
                        }
                    }
                }
            }
        },
...
        "/api/books/{bookID}/prices/{priceID}": {
            "delete": {
                "tags": [
                    "prices"
                ],
                "summary": "Delete a manual rate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "price id",
                        "name": "priceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
...
        "routes.AccessRequestDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "decided_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routes.adminUserRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.inviteRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}